@@ -31,33 +31,28 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  hive: ^2.2.3
-  hive_flutter: ^1.1.0
+  hive_ce: ^2.11.3
+  hive_ce_flutter: ^2.3.1
   path_provider: ^2.0.11
-  geolocator: ^11.0.0
-  geocoding: ^3.0.0
-  permission_handler: ^11.0.0
+  geolocator: ^14.0.2
+  geocoding: ^4.0.0
+  permission_handler: ^12.0.1
   image_picker: ^1.0.7
   path: ^1.8.3
   logging: ^1.2.0
-  build_runner: ^2.4.0
-  mockito: ^5.4.0
   path_provider_platform_interface: ^2.1.0
   plugin_platform_interface: ^2.1.0
   provider: ^6.1.5
   collection: ^1.17.0
+  uuid: ^4.5.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  hive_generator: ^2.0.0
-
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
-  flutter_lints: ^5.0.0
+  hive_ce_generator: ^1.9.3
+  flutter_lints: ^6.0.0
+  build_runner: ^2.6.0
+  mockito: ^5.5.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
