@@ -1,5 +1,5 @@
 // lib/models/location_model.dart
-import 'package:hive/hive.dart';
+import 'package:hive_ce/hive.dart';
 
 part 'location_model.g.dart'; // Will be generated by build_runner for Hive
 
@@ -18,29 +18,33 @@ class Location extends HiveObject {
   String? address;
 
   @HiveField(5)
-  List<String> imagePaths;
+  List<String>? imageGuids;
 
   Location({
     required this.id,
     required this.name,
     this.description,
     this.address,
-    List<String>? imagePaths,
-  }) : imagePaths = imagePaths ?? [];
+    List<String>? imageGuids,
+  }) : imageGuids = imageGuids ?? [];
 
   Location copyWith({
     String? id,
     String? name,
     String? description,
     String? address,
-    List<String>? imagePaths,
+    List<String>? imageGuids,
   }) {
     return Location(
       id: id ?? this.id,
       name: name ?? this.name,
       description: description ?? this.description,
       address: address ?? this.address,
-      imagePaths: imagePaths ?? List.from(this.imagePaths),
+      imageGuids: imageGuids != null
+          ? List<String>.from(imageGuids)
+          : (this.imageGuids != null
+                ? List<String>.from(this.imageGuids!)
+                : []),
     );
   }
 }
